+++ backend/src/test/java/com/mpezzanera/app/whather_app/controller/CityControllerTest.java
@@ -12,13 +12,16 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.context.annotation.Import;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 
+import com.mpezzanera.app.whather_app.config.SecurityConfig;
 import com.mpezzanera.app.whather_app.model.City;
 import com.mpezzanera.app.whather_app.service.OpenWeatherMapService;
 
 @WebMvcTest(CityController.class)
+@Import(SecurityConfig.class)
 class CityControllerTest {
 
     @Autowired